# authapi/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Permission Management API documentation",
    "description" : "This is an API for managing application permissions.",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/polarfox/AuthAPI/1.0.0"
  } ],
  "security" : [ {
    "implicit" : [ "read", "write" ]
  } ],
...
  "paths" : {
...
    "/checkpermission" : {
      "get" : {
        "tags" : [ "authorizationAPI" ],
        "summary" : "Check whether Permission is granted for user",
        "description" : "Check whether Permission is granted for user",
        "operationId" : "checkPermission",
        "parameters" : [ {
          "name" : "entity",
          "in" : "query",
          "description" : "The entity name",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "DataProduct"
        }, {
          "name" : "action",
          "in" : "query",
          "description" : "The action name",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "Create"
        }, {
          "name" : "entityId",
          "in" : "query",
          "description" : "The particular entity Id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Permission is granted to user"
          },
          "400" : {
            "description" : "Invalid syntax",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Invalid query parameter 'role'. Valid values are 'roleId'"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "401" : {
            "description" : "Permission not granted to user"
          },
          "405" : {
            "description" : "Requested method not allowed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Method 'Put' not allowed. Allowed methods are 'Get' and 'Post'."
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Internal server error"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        }
      }
    },
    "/checkrole" : {
      "get" : {
        "tags" : [ "authorizationAPI" ],
        "summary" : "Check whether Role is granted for user",
        "description" : "Check whether Role is granted for user",
        "operationId" : "checkRole",
        "parameters" : [ {
          "name" : "roleId",
          "in" : "query",
          "description" : "The Role Id",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "description" : "The User ID",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Role is granted to user"
          },
          "400" : {
            "description" : "Invalid syntax",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Invalid query parameter 'role'. Valid values are 'roleId'"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "401" : {
            "description" : "Role not granted to user"
          },
          "405" : {
            "description" : "Requested method not allowed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Method 'Put' not allowed. Allowed methods are 'Get' and 'Post'."
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Internal server error"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        }
      }
    },
    "/checkrole/{roleId}" : {
      "get" : {
        "tags" : [ "authorizationAPI" ],
        "summary" : "Check whether Role is granted for current user",
        "description" : "Check whether Role is granted for current user",
        "operationId" : "checkRoleById",
        "parameters" : [ {
          "name" : "roleId",
          "in" : "path",
          "description" : "The Role Id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Role is granted to user"
          },
          "400" : {
            "description" : "Invalid syntax",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Invalid query parameter 'role'. Valid values are 'roleId'"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "401" : {
            "description" : "Role not granted to user"
          },
          "405" : {
            "description" : "Requested method not allowed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Method 'Put' not allowed. Allowed methods are 'Get' and 'Post'."
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                },
                "example" : {
                  "errors" : [ {
                    "message" : "Internal server error"
                  } ],
                  "reqId" : "123e4567-e89b-12d3-a456-426614174000"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Error" : {
        "required" : [ "errors" ],
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "Array of detailed error messages",
            "items" : {
              "$ref" : "#/components/schemas/Error_errors"
            }
          },
          "reqId" : {
            "type" : "string",
            "description" : "Server generated UUID for the request to help with API issue solving"
          }
        }
      },
...
      "Error_errors" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "example" : null
      }
    },
...
    "securitySchemes" : {
      "implicit" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://example.com/oauth/auth",
            "scopes" : {
              "write" : "allows modifying resources",
              "read" : "allows reading resources"
            }
          }
        }
      }
    }
  }
}